update: add standalone tensor dump file

+++ src/type.h
@@ -15,7 +15,7 @@ using u64 = uint64_t;
 using f32 = float;
 using f64 = double;
 using str = std::string;
-using token_id = f32;
+using token_id = u32;
 using token_type = i32;
 
 #endif //GEMMA_GGML_TYPE_H
